fix(General/UI): removed editor platform from front end ASMDEF and changed convert to int32 to double for priority

+++ Assets/Scripts/Frontend/CreateStruct.cs
@@ -86,7 +86,7 @@ public class CreateStruct : MonoBehaviour
                 }
                 // If value is 0, then no priority was selected
 
-                double prioNum = Convert.ToUInt32(priorityText.text);
+                double prioNum = Convert.ToDouble(priorityText.text);
                 
                 StaticData.priority[StaticData.arrIndex] = (direction,prioNum); 
                 StaticData.arrIndex++;
